requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
borb==2.1.12
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
config==0.5.1
cryptography==40.0.2
fonttools==4.39.3
frozenlist==1.3.3
idna==3.4
lxml==4.9.2
multidict==6.0.4
openai==0.27.6
Pillow==9.5.0
pycparser==2.21
pypng==0.20220715.0
python-barcode==0.14.0
python-decouple==3.8
qrcode==7.4.2
requests==2.30.0
tqdm==4.65.0
typing_extensions==4.5.0
urllib3==2.0.2
yarl==1.9.2
